+++ shared/CHANGELOG.md
@@ -11,7 +11,7 @@
 - [23. 監督・コーチの出場履歴(Manager-Appearance)]
 - [24. 選手の試合イベントログ(Player-Match-Event-Log)]
 - [25. 監督・コーチの試合イベントログ(Staff-Match-Event-Log)]
-- [26. 試合でのフォーメーション(Match-Team-Formation)]
+- [26. 試合でのフォーメーション(Team-Match-Formation)]
 
 ### v1.0.10